# src/IO.Swagger/Models/RecipesList.cs
/*
 * API de recettes
 *
 * Une API pour créer, lire, mettre à jour et supprimer des recettes.
 *
 * OpenAPI spec version: 1.0.0
 * 
 * Generated by: https://github.com/swagger-api/swagger-codegen.git
 */
using System;
using System.Linq;
using System.IO;
using System.Text;
using System.Collections;
using System.Collections.Generic;
using System.Collections.ObjectModel;
using System.ComponentModel.DataAnnotations;
using System.Runtime.Serialization;
using Newtonsoft.Json;

namespace IO.Swagger.Models
{
    /// <summary>
    /// 
    /// </summary>
    [DataContract]
    public partial class RecipesList : IEquatable<RecipesList>
    { 
        /// <summary>
        /// Gets or Sets Recipes
        /// </summary>

        [DataMember(Name="recipes")]
        public List<Recipe> Recipes { get; set; }

        /// <summary>
        /// Returns the string presentation of the object
        /// </summary>
        /// <returns>String presentation of the object</returns>
        public override string ToString()
        {
            var sb = new StringBuilder();
            sb.Append("class RecipesList {\n");
            sb.Append("  Recipes: ").Append(Recipes).Append("\n");
            sb.Append("}\n");
            return sb.ToString();
        }

        /// <summary>
        /// Returns the JSON string presentation of the object
        /// </summary>
        /// <returns>JSON string presentation of the object</returns>
        public string ToJson()
        {
            return JsonConvert.SerializeObject(this, Formatting.Indented);
        }

        /// <summary>
        /// Returns true if objects are equal
        /// </summary>
        /// <param name="obj">Object to be compared</param>
        /// <returns>Boolean</returns>
        public override bool Equals(object obj)
        {
            if (ReferenceEquals(null, obj)) return false;
            if (ReferenceEquals(this, obj)) return true;
            return obj.GetType() == GetType() && Equals((RecipesList)obj);
        }

        /// <summary>
        /// Returns true if RecipesList instances are equal
        /// </summary>
        /// <param name="other">Instance of RecipesList to be compared</param>
        /// <returns>Boolean</returns>
        public bool Equals(RecipesList other)
        {
            if (ReferenceEquals(null, other)) return false;
            if (ReferenceEquals(this, other)) return true;

            return 
                (
                    Recipes == other.Recipes ||
                    Recipes != null &&
                    Recipes.SequenceEqual(other.Recipes)
                );
        }

        /// <summary>
        /// Gets the hash code
        /// </summary>
        /// <returns>Hash code</returns>
        public override int GetHashCode()
        {
            unchecked // Overflow is fine, just wrap
            {
                var hashCode = 41;
                // Suitable nullity checks etc, of course :)
                    if (Recipes != null)
                    hashCode = hashCode * 59 + Recipes.GetHashCode();
                return hashCode;
            }
        }

        #region Operators
        #pragma warning disable 1591

        public static bool operator ==(RecipesList left, RecipesList right)
        {
            return Equals(left, right);
        }

        public static bool operator !=(RecipesList left, RecipesList right)
        {
            return !Equals(left, right);
        }

        #pragma warning restore 1591
        #endregion Operators
    }
}